# api/openapi.yaml
openapi: 3.1.0
info:
  title: PACEFOLIO API
  version: 0.2.0-draft
  description: >
    API 계약 (R2 B3b 재작성 · R4 보안 계약 · R7 P0-5 구현 정합). 도메인 모델 = packages/domain, 규칙 = docs/02~09.
    모든 요청은 서버가 (1)역할 능력 canAny() (2)테넌트/소유/담당 스코프(authorization.ts)를 검증.
    클라이언트가 준 role·academyId·participantId·otpVerified 등은 절대 신뢰하지 않는다.
    enum 값은 packages/domain/enums.ts 와 x-domain-enum 마커로 자동 대조된다(test/openapi-drift.test.ts).

    [구현 현황 — 정본 = apps/api/src/app.ts]
    구현된 경로(auth start/callback·dev login·sessions me/logout/logout-all·
    guardian-links·payments/prepare·invoices·webhooks/pg)는 실제 응답 형식·상태코드
    기준으로 기술한다. "(계약 초안 — 미구현)" 표기 경로는 착공 전 계약이며 구현 시
    이 문서를 구현 기준으로 갱신한다. 오류 응답 형식은 components/schemas/Error 참조
    ({ error: "CODE" } — requestId 필드는 추후 공통 오류 계약(P1)에서 추가 예정).

    [CSRF 계약 — R4 §11 · 상세 docs/11-security-contracts.md]
    상태를 변경하는 모든 요청(POST/PUT/PATCH/DELETE — 결제준비·환불요청/승인·
    동의변경·탈퇴·로그아웃·전기기 로그아웃·SupportView 발급/철회)은 다음을 모두 만족:
    (1) 세션 쿠키 SameSite=Lax (2) 서버 Origin 검증(allowlist 불일치 = 403)
    (3) X-CSRF-Token 헤더 = double-submit cookie 검증.
    Admin 은 별도 도메인/앱으로 분리하며 독립 CSRF 방어를 갖는다.
servers:
  - url: https://api.pacefolio.example/v1
security:
  - sessionCookie: []
...
paths:
...
  /auth/{provider}/start:
    post:
      tags: [auth]
      security: []
      summary: 소셜 로그인 시작 — 카카오(앵커)·네이버·구글·애플
      description: >
        [OAuth 보안 계약 — R4 §12] state 는 서버 저장·일회성 소비·10분 만료.
        PKCE(S256) code_challenge 필수. OIDC provider(구글·애플)는 nonce 포함.
        redirect URI 는 서버 allowlist 로만. authorization code 는 1회 사용 후 폐기.
        계정 병합 금지: 동일 이메일이라도 서로 다른 소셜 계정을 자동 병합하지
        않는다 — 병합은 별도 인증 절차 + 사용자 명시 확인 필요.
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string, enum: [kakao, naver, google, apple] }
      responses:
        '200': { description: 'authorize URL + state(서버 저장·일회성·만료)', content: { application/json: { schema: { type: object, properties: { authorizeUrl: { type: string }, state: { type: string } } } } } }
        '404': { description: UNKNOWN_PROVIDER — kakao/naver/google/apple 외, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '501': { description: PROVIDER_NOT_CONFIGURED — 키 미설정 provider, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /auth/{provider}/callback:
    get:
      tags: [auth]
      security: []
      summary: OAuth callback — 세션 발급. 신규 사용자는 약관·동의 단계로.
      description: >
        서버 검증 순서: state 존재·미사용·미만료 확인 후 즉시 소비 →
        PKCE code_verifier 대조 → code 교환(1회) → OIDC nonce 대조 →
        provider 의 verified email/phone 만 신뢰. 어느 하나 실패 = 401(로그인 CSRF 방어).
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string }
        - { name: code, in: query, required: true, schema: { type: string } }
        - { name: state, in: query, required: true, schema: { type: string } }
      responses:
        '200': { description: '세션 쿠키 설정(HttpOnly·Secure·SameSite=Lax·rotation). body = { userId, isNewUser } — 세션 정보는 쿠키+GET /sessions/me 로', content: { application/json: { schema: { type: object, required: [userId, isNewUser], properties: { userId: { type: string }, isNewUser: { type: boolean } } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '501': { description: PROVIDER_NOT_CONFIGURED, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /auth/dev/login:
    post:
      tags: [auth]
      security: []
      summary: 개발용 로그인 — 카카오 키 없이 세션 발급 (dev 전용, 프로덕션 404)
      description: >
        게이트: enableDevLogin && NODE_ENV !== production — 미충족 시 404 NOT_FOUND
        (존재 자체를 숨김, PG_SIMULATION 과 같은 패턴). 이름 기반 seed 사용자를
        재사용해 결정적 userId 로 세션 쿠키를 발급한다. 실 provider 연동과 무관.
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [name], properties: { name: { type: string, minLength: 1, maxLength: 30 } }, additionalProperties: false } } }
      responses:
        '200': { description: 세션 쿠키 설정 + userId, content: { application/json: { schema: { type: object, required: [userId], properties: { userId: { type: string } } } } } }
        '404': { description: 프로덕션 또는 dev login 비활성 — NOT_FOUND, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '422': { description: INVALID_BODY, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /sessions/me:
    get:
      tags: [auth]
      summary: 현재 세션·소속·역할 — route guard 의 진실. 멤버십은 서버 도출.
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Session' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /academies/{academyId}/guardian-links:
    post:
      tags: [linking]
      summary: >
        보호자-자녀 연결 — 서버 OTP 증적 + 학원 선등록 보호자 연락처(또는 초대코드) 결합.
        이름·생년 일치만으로 자동 VERIFIED 금지(guardian-linking.ts evaluateLink 재현).
      description: >
        CSRF 계약 적용(Origin + X-CSRF-Token). actorUserId 는 세션에서 도출 — body 로 받지 않음.
        body 는 strict — 예상하지 않은 필드는 422 INVALID_BODY(R5 §6.2).
        원자 트랜잭션: OTP 검증세션 1회 소비 → 초대코드/선등록 연락처 결합 → 링크 생성.
        동시 요청은 DB UNIQUE·FOR UPDATE 로 직렬화 — 경쟁 패자는 409 CONFLICT.
      parameters:
        - $ref: '#/components/parameters/AcademyId'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/GuardianLinkRequest' } } }
      responses:
        '201': { description: 'VERIFIED — { status, linkId, participantId }', content: { application/json: { schema: { $ref: '#/components/schemas/LinkResult' } } } }
        '202': { description: 'PENDING(수동 심사)·REJECTED — { status, reason }', content: { application/json: { schema: { $ref: '#/components/schemas/LinkResult' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: 'CONFLICT — OTP 동시 소비 경쟁 패자·중복 링크(UNIQUE 위반)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '422': { description: 'INVALID_BODY(+issues 필드 경로 배열)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ── 출결 (예정 ≠ 실제) ────────────────────────────────────────
...
  /academies/{academyId}/invoices:
    get:
      tags: [billing]
      summary: 내 자녀 청구서 목록 — 보호자 관점(VERIFIED 연결 자녀분만)
      description: >
        구현 기준 응답 = { invoices: GuardianInvoice[] } (billing/queries.ts listGuardianInvoices).
        페이지네이션 미구현 — 현재는 전량 반환. page[size]/page[cursor] 커서 계약은
        추후 목록 API 공통화(P1) 때 적용. 원장 관점(학원 전체) 목록도 미구현.
      parameters:
        - $ref: '#/components/parameters/AcademyId'
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, required: [invoices], properties: { invoices: { type: array, items: { $ref: '#/components/schemas/GuardianInvoice' } } } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /academies/{academyId}/payments/prepare:
    post:
      tags: [billing]
      summary: >
        합산 결제 준비 — 서버가 invoiceIds 로 금액 재계산·allocation 생성.
        같은 학원 Invoice 만 혼합 가능(canGuardianPayInvoices). GUARDIAN 역할 필수.
      description: >
        멱등 의미(idempotency.ts, 구현 기준): 같은 key+body → 200 재생 / 다른 body →
        409 IDEMPOTENCY_KEY_REUSED / 처리중 → 409 IN_PROGRESS (202 아님).
        R7 P0-3: 대상 Invoice 에 활성 attempt(AUTHORIZED 또는 미만료 PENDING)가 있으면
        멱등키가 달라도 409 ACTIVE_PAYMENT_ATTEMPT_EXISTS(+paymentId).
        UI 성공 ≠ PG CAPTURED — 확정은 webhook 만. CSRF 계약 적용.
      parameters:
        - $ref: '#/components/parameters/AcademyId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PaymentInput' } } }
      responses:
        '201': { description: '생성됨 — { paymentId, amount, status: PENDING }', content: { application/json: { schema: { $ref: '#/components/schemas/PaymentPrepareResult' } } } }
        '200': { description: '멱등 재생(같은 key+body) — 기존 결제 반환', content: { application/json: { schema: { $ref: '#/components/schemas/PaymentPrepareResult' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: 'error = IN_PROGRESS | IDEMPOTENCY_KEY_REUSED | ACTIVE_PAYMENT_ATTEMPT_EXISTS(+paymentId)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '422': { description: 'error = INVALID_BODY | IDEMPOTENCY_KEY_REQUIRED | UNPROCESSABLE(+reason — 권한/정산 거부 사유)', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /webhooks/pg/{provider}:
    post:
      tags: [webhooks]
      security: []
      summary: >
        PG 웹훅 수신 — 처리 순서: 인증(fail-closed) → inbox unique insert(중복 차단) →
        webhooks.ts decidePaymentWebhook(중복/역순/재조회 판단) →
        상태변경+outbox 를 한 트랜잭션으로. 오래된 이벤트가 최종 상태를 되돌리지 않는다.
      description: >
        [인증 — R7 P0-1 fail-closed] provider = mockpg(개발 시뮬 전용):
        enableMockPg && NODE_ENV !== production && 시크릿 설정 시에만 열림(아니면 404),
        X-Webhook-Secret 헤더 불일치 = 401 SIGNATURE_INVALID.
        그 외 provider: 등록된 verifier 가 raw body 서명을 검증해야 통과(불일치 401),
        미등록 provider = 404 NOT_FOUND(allowlist).

        ⚠️ 현재 requestBody(paymentId·targetStatus 직접 지정)는 mockpg 전용(dev).
        실 provider adapter 는 미구현 — 연동 시 adapter 가 서명 검증 후
        (provider, providerPaymentId)로 내부 결제를 찾고 provider event type 을
        내부 상태로 매핑한다(R7 P0-4). 그때 이 스키마를 provider event 계약으로 대체.
      parameters:
        - { name: provider, in: path, required: true, schema: { type: string } }
        - { name: X-Webhook-Secret, in: header, required: false, schema: { type: string }, description: 'mockpg 전용 공유 시크릿. 실 provider 는 verifier 가 자체 서명 헤더를 검증' }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PgWebhookEvent' } } }
      responses:
        '200': { description: '수신(중복·stale 포함 — 항상 200, 내부 판단은 inbox). body = { decision: APPLY | IGNORE_ALREADY_SEEN | ACK_NO_STATE_CHANGE | IGNORE_STALE | RECONCILE | REJECT_INVALID }', content: { application/json: { schema: { type: object, required: [decision], properties: { decision: { type: string } } } } } }
        '401': { description: 'SIGNATURE_INVALID — 서명/시크릿 불일치', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: 'NOT_FOUND — 미등록 provider 또는 mockpg 게이트 미충족', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '422': { description: INVALID_BODY, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ── 사진 동의 (P0-9) ─────────────────────────────────────────
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pf_session
      description: >
        [R4 §11] 속성: HttpOnly · Secure · SameSite=Lax · Path=/ ·
        Max-Age=세션정책(권한 상승·로그인 시 rotation). 상태 변경 요청은
        info.description 의 CSRF 계약(Origin 검증 + X-CSRF-Token)을 함께 만족해야 함.
  parameters:
    AcademyId:
      name: academyId
      in: path
      required: true
      schema: { type: string }
      description: 테넌트 격리 축. 세션 소속과 불일치 시 403 TENANT_SCOPE.
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string }
      description: >
        멱등 의미(idempotency.ts, 구현 기준): (academy, actor, operation, key, bodyHash) 기준.
        같은 key+body = 200 응답 재생 / 다른 body = 409 IDEMPOTENCY_KEY_REUSED /
        처리중 = 409 IN_PROGRESS / 보관기간(24h) 만료 후 = 신규.
        헤더 누락·128자 초과 = 422 IDEMPOTENCY_KEY_REQUIRED.
...
  responses:
    Unauthorized: { description: 미인증/세션 만료, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Forbidden: { description: 권한/테넌트/담당/연결 스코프 위반, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    Error:
      type: object
      description: >
        구현 기준 오류 형식 = { error: "CODE" } + 상황별 선택 필드(reason·paymentId·issues).
        code 예: UNAUTHORIZED · FORBIDDEN_ORIGIN · CSRF_TOKEN_MISMATCH ·
        FORBIDDEN_TENANT_SCOPE · MEMBERSHIP_NOT_ACTIVE · FORBIDDEN_ROLE ·
        PLATFORM_ADMIN_SEPARATE_BOUNDARY · NOT_FOUND · UNKNOWN_PROVIDER ·
        PROVIDER_NOT_CONFIGURED · CONFLICT · IN_PROGRESS · IDEMPOTENCY_KEY_REUSED ·
        ACTIVE_PAYMENT_ATTEMPT_EXISTS · IDEMPOTENCY_KEY_REQUIRED · INVALID_BODY ·
        UNPROCESSABLE · SIGNATURE_INVALID.
        requestId 필드는 추후 공통 오류 계약(P1)에서 추가 예정.
      required: [error]
      properties:
        error: { type: string, description: 오류 코드 — 위 description 의 목록 참조 }
        reason: { type: string, description: 사람이 읽는 거부 사유(422 UNPROCESSABLE 등) }
        paymentId: { type: string, description: ACTIVE_PAYMENT_ATTEMPT_EXISTS 시 진행 중 결제 ID }
        issues: { type: array, items: { type: string }, description: INVALID_BODY 시 zod 검증 실패 필드 경로 }
    Session:
      type: object
      description: GET /sessions/me 응답 — sessions/service.ts ResolvedSession 과 1:1.
      required: [user, memberships]
      properties:
        user:
          type: object
          required: [id, name, phone, email]
          properties:
            id: { type: string }
            name: { type: string }
            phone: { type: string }
            email: { type: ['string', 'null'] }
        memberships:
          type: array
          description: 서버 도출 — 클라 입력 아님. 멀티역할 모델 A(roles[]).
          items:
            type: object
            properties:
              academyId: { type: string }
              roles: { type: array, items: { type: string, x-domain-enum: ROLES, enum: [OWNER, MANAGER, COACH, DESK, DRIVER, GUARDIAN, PLATFORM_ADMIN] } }
              status: { type: string, x-domain-enum: MEMBERSHIP_STATUS, enum: [INVITED, ACTIVE, SUSPENDED, ENDED] }
    GuardianLinkRequest:
      type: object
      description: >
        apps/api LinkBody(zod strict)와 1:1 — 예상하지 않은 필드는 422. otpVerified boolean 없음.
        academyId 는 body 가 아니라 path 로 받는다. actorUserId 는 세션 도출.
      required: [verificationSessionId, childName, childBirth, relationshipType, consentPolicyVersion, consentAgreed]
      additionalProperties: false
      properties:
        verificationSessionId: { type: string, description: OTP 검증이 발급한 서버 증적(GuardianVerificationSession) }
        childName: { type: string }
        childBirth: { type: string, format: date }
        relationshipType: { type: string, x-domain-enum: RELATIONSHIP_TYPE, enum: [MOTHER, FATHER, GRANDPARENT, LEGAL_GUARDIAN, OTHER] }
        consentPolicyVersion: { type: string }
        consentAgreed: { type: boolean }
        academyInviteCode: { type: string, description: 선등록 연락처 불일치 시 대안 결합 근거 }
    LinkResult:
      type: object
      description: >
        구현 기준(linking/service.ts LinkSliceResult): VERIFIED(201) 은 linkId·participantId 포함,
        PENDING·REJECTED(202) 는 reason 포함. UNVERIFIED 는 응답으로 쓰이지 않음(도메인 enum 전체 대조용).
      required: [status]
      properties:
        status: { type: string, x-domain-enum: VERIFICATION_STATUS, enum: [UNVERIFIED, PENDING, VERIFIED, REJECTED] }
        linkId: { type: string }
        participantId: { type: string }
        reason: { type: string }
...
    GuardianInvoice:
      type: object
      description: >
        GET /academies/{academyId}/invoices 응답 요소 — billing/queries.ts GuardianInvoiceRow 와 1:1.
        netPaidAmount·outstandingAmount 도출은 서버 내부(billing.ts) — 목록 응답에는 미포함(추후 확장).
      required: [invoiceId, participantId, participantName, status, total, dueDate, lines]
      properties:
        invoiceId: { type: string }
        participantId: { type: string }
        participantName: { type: string }
        status: { type: string, x-domain-enum: INVOICE_STATUS, enum: [DRAFT, ISSUED, PARTIALLY_PAID, PAID, OVERDUE, VOID, REFUNDED] }
        total: { type: integer, description: KRW. = Σ lines }
        dueDate: { type: string, format: date }
        lines:
          type: array
          items:
            type: object
            required: [type, label, amount]
            properties:
              type: { type: string, description: TUITION·TRANSPORT 등 InvoiceLine 유형 }
              label: { type: string }
              amount: { type: integer }
    PaymentInput:
      type: object
      description: >
        apps/api PrepareBody(zod strict)와 1:1 — 예상하지 않은 필드는 422 INVALID_BODY.
        금액은 서버가 invoiceIds 로 전액 재계산 — 클라 금액 필드는 받지 않는다.
      required: [invoiceIds]
      additionalProperties: false
      properties:
        invoiceIds:
          type: array
          minItems: 1
          maxItems: 10
          items: { type: string }
          description: 같은 학원 소속 + 결제 가능(canPay) 자녀분만. 위반 시 422 UNPROCESSABLE(+reason).
    PaymentPrepareResult:
      type: object
      description: POST payments/prepare 201(생성)·200(멱등 재생) 공통 응답.
      required: [paymentId, amount, status]
      properties:
        paymentId: { type: string }
        amount: { type: integer, description: 서버 재계산 합산 금액(KRW) }
        status: { type: string, description: 'Payment 상태 — 생성 직후는 PENDING, 재생은 저장된 현재 상태(PAYMENT_STATUS enum)' }
...
    PgWebhookEvent:
      type: object
      description: >
        ⚠️ mockpg(개발 시뮬) 전용 body — apps/api WebhookBody(zod strict)와 1:1.
        paymentId·targetStatus 를 직접 지정하는 형식은 dev 에서만 유효.
        실 provider 연동 시 adapter 가 서명 검증 후 provider event
        (providerPaymentId·event type)를 내부 결제·상태로 매핑한다(R7 P0-4 — 미구현).
      required: [providerEventId, paymentId, targetStatus, occurredAt]
      additionalProperties: false
      properties:
        providerEventId: { type: string, description: '(provider, providerEventId) unique — inbox 중복 방지 축' }
        paymentId: { type: string, description: 내부 Payment ID (mockpg 전용 — 실 provider 는 providerPaymentId 매핑) }
        targetStatus: { type: string, x-domain-enum: PAYMENT_STATUS, enum: [PENDING, AUTHORIZED, CAPTURED, FAILED, CANCELLED, PARTIALLY_REFUNDED, REFUNDED] }
        occurredAt: { type: string, format: date-time, description: PG 발생시각 — monotonic guard 기준 }
...